# specs/001-interactive-teaching/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Interactive Teaching API
  version: 1.1.0
  description: REST API for lessons and interactive modal content. UTF-8 JSON.

servers:
  - url: http://localhost:8000
    description: Local development

paths:
  /health:
    get:
      summary: Liveness probe
      operationId: healthCheck
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [ok]

  /lessons/{lesson_id}:
    get:
      summary: Get lesson with TOC, multimedia strip, and TipTap JSON content
      operationId: getLesson
      parameters:
        - name: lesson_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lesson document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "404":
          description: Not found

    put:
      summary: Replace lesson
      operationId: putLesson
      parameters:
        - name: lesson_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonInput"
      responses:
        "200":
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"

    patch:
      summary: Partially update lesson (e.g. multimedia_strip only)
      operationId: patchLesson
      parameters:
        - name: lesson_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonPatch"
      responses:
        "200":
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"

  /interactive-nodes:
    post:
      summary: Create interactive node (multimedia / text)
      operationId: createInteractiveNode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InteractiveNodeCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: string
        "422":
          description: Validation error

  /interactive-nodes/{node_id}:
    get:
      summary: Resolve interactive node for modal rendering
      operationId: getInteractiveNode
      parameters:
        - name: node_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resolved payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InteractiveNodeResolved"
        "404":
          description: Unknown node
        "422":
          description: URL not allowed or invalid payload

    patch:
      summary: Update interactive node (partial; source merge for nested keys)
      operationId: patchInteractiveNode
      parameters:
        - name: node_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InteractiveNodePatch"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: Unknown node

components:
  schemas:
    TOCItem:
      type: object
      required: [id, label]
      properties:
        id:
          type: string
        label:
          type: string
        children:
          type: array
          items:
            $ref: "#/components/schemas/TOCItem"

    MultimediaStripItem:
      type: object
      required: [node_id]
      properties:
        node_id:
          type: string
        label:
          type: string
        kind:
          type: string
          description: Populated on GET for chip UI (from interactive_nodes)

    MultimediaStripItemInput:
      type: object
      required: [node_id]
      properties:
        node_id:
          type: string
        label:
          type: string

    Lesson:
      type: object
      required: [id, title, toc, content]
      properties:
        id:
          type: string
        title:
          type: string
        toc:
          type: array
          items:
            $ref: "#/components/schemas/TOCItem"
        content:
          description: TipTap / ProseMirror JSON document
          type: object
          additionalProperties: true
        multimedia_strip:
          type: array
          items:
            $ref: "#/components/schemas/MultimediaStripItem"
        updated_at:
          type: string
          format: date-time

    LessonInput:
      type: object
      required: [title, toc, content]
      properties:
        title:
          type: string
        toc:
          type: array
          items:
            $ref: "#/components/schemas/TOCItem"
        content:
          type: object
          additionalProperties: true
        multimedia_strip:
          type: array
          items:
            $ref: "#/components/schemas/MultimediaStripItemInput"

    LessonPatch:
      type: object
      properties:
        title:
          type: string
        toc:
          type: array
          items:
            $ref: "#/components/schemas/TOCItem"
        content:
          type: object
          additionalProperties: true
        multimedia_strip:
          type: array
          items:
            $ref: "#/components/schemas/MultimediaStripItemInput"

    InteractiveNodeCreate:
      type: object
      required: [kind, label, source]
      properties:
        kind:
          type: string
          enum: [text, image, audio, video, web]
        label:
          type: string
        source:
          type: object
          additionalProperties: true
        alt:
          type: string
        embed:
          type: string
        embed_id:
          type: string

    InteractiveNodePatch:
      type: object
      properties:
        label:
          type: string
        source:
          type: object
          additionalProperties: true
        alt:
          type: string
        embed:
          type: string
        embed_id:
          type: string

    InteractiveNodeResolved:
      oneOf:
        - $ref: "#/components/schemas/ResolvedText"
        - $ref: "#/components/schemas/ResolvedImage"
        - $ref: "#/components/schemas/ResolvedAudio"
        - $ref: "#/components/schemas/ResolvedVideo"
        - $ref: "#/components/schemas/ResolvedWeb"

    ResolvedText:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          enum: [text]
        text:
          type: string
          description: Legacy plain string (optional when content is set)
        content:
          type: object
          description: TipTap JSON document
          additionalProperties: true

    ResolvedImage:
      type: object
      required: [kind, url]
      properties:
        kind:
          type: string
          enum: [image]
        url:
          type: string
          format: uri
        alt:
          type: string

    ResolvedAudio:
      type: object
      required: [kind, url]
      properties:
        kind:
          type: string
          enum: [audio]
        url:
          type: string
          format: uri

    ResolvedVideo:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          enum: [video]
        url:
          type: string
          format: uri
        embed:
          type: string
          description: e.g. youtube
        embed_id:
          type: string

    ResolvedWeb:
      type: object
      required: [kind, url]
      properties:
        kind:
          type: string
          enum: [web]
        url:
          type: string
          format: uri